Don't compile in VM-specific things in toit.compile
