+++ src/vm.cc
@@ -43,7 +43,6 @@ VM::VM() {
   OS::reset_monotonic_time();  // Reset "up time".
   Primitive::set_up();
   _heap_memory = _new HeapMemory();
-  _program_heap_memory = _new ProgramHeapMemory();
   _scheduler = _new Scheduler();
 
   _event_manager = _new EventSourceManager();
@@ -55,7 +54,6 @@ VM::~VM() {
   delete _event_manager;
   delete _scheduler;
   delete _heap_memory;
-  delete _program_heap_memory;
   _current = null;
 }
